refactor: try to move to one goldenfile
