--- src/query/service/tests/it/storages/system/clusters_table.rs
@@ -1,35 +1,0 @@
-// Copyright 2021 Datafuse Labs.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-use common_base::base::tokio;
-use common_exception::Result;
-use databend_query::storages::system::ClustersTable;
-use databend_query::storages::TableStreamReadWrap;
-use databend_query::storages::ToReadDataSourcePlan;
-use futures::TryStreamExt;
-
-#[tokio::test(flavor = "multi_thread", worker_threads = 1)]
-async fn test_clusters_table() -> Result<()> {
-    let (_guard, ctx) = crate::tests::create_query_context().await?;
-    let table = ClustersTable::create(1);
-
-    let source_plan = table.read_plan(ctx.clone(), None).await?;
-
-    let stream = table.read(ctx, &source_plan).await?;
-    let result = stream.try_collect::<Vec<_>>().await?;
-    let block = &result[0];
-    assert_eq!(block.num_columns(), 4);
-
-    Ok(())
-}
