--- src/query/service/tests/it/storages/system/testdata/tables.txt
@@ -1,40 +1,0 @@
----------- TABLE INFO ------------
-DB.Table: 'system'.'clusters', Table: clusters-table_id:1, ver:0, Engine: SystemClusters
--------- TABLE CONTENTS ----------
-+------------------------+-----------+------+
-| name                   | host      | port |
-+------------------------+-----------+------+
-| Ib6r09zSYQlrhNPZq2Zl56 | 127.0.0.1 | 9090 |
-+------------------------+-----------+------+
-
----------- TABLE INFO ------------
-DB.Table: 'system'.'settings', Table: settings-table_id:1, ver:0, Engine: SystemSettings
--------- TABLE CONTENTS ----------
-+--------------------------------+------------+------------+---------+----------------------------------------------------------------------------------------------------+--------+
-| name                           | value      | default    | level   | description                                                                                        | type   |
-+--------------------------------+------------+------------+---------+----------------------------------------------------------------------------------------------------+--------+
-| enable_async_insert            | 0          | 0          | SESSION | Whether the client open async insert mode, default value: 0                                        | UInt64 |
-| enable_cbo                     | 1          | 1          | SESSION | If enable cost based optimization, default value: 1                                                | UInt64 |
-| enable_new_processor_framework | 1          | 1          | SESSION | Enable new processor framework if value != 0, default value: 1                                     | UInt64 |
-| enable_planner_v2              | 1          | 1          | SESSION | Enable planner v2 by setting this variable to 1, default value: 1                                  | UInt64 |
-| flight_client_timeout          | 60         | 60         | SESSION | Max duration the flight client request is allowed to take in seconds. By default, it is 60 seconds | UInt64 |
-| format_compression             | None       | None       | SESSION | Format compression, default value: None                                                            | String |
-| format_empty_as_default        | 1          | 1          | SESSION | Format empty_as_default, default value: 1                                                          | UInt64 |
-| format_field_delimiter         | ,          | ,          | SESSION | Format field delimiter, default value: ,                                                           | String |
-| format_quote_char              | '"'        | '"'        | SESSION | The quote char for CSV. default value: '"'.                                                        | String |
-| format_record_delimiter        | "\n"       | "\n"       | SESSION | Format record_delimiter, default value: "\n"                                                       | String |
-| format_skip_header             | 0          | 0          | SESSION | Whether to skip the input header, default value: 0                                                 | UInt64 |
-| group_by_two_level_threshold   | 10000      | 10000      | SESSION | The threshold of keys to open two-level aggregation, default value: 10000                          | UInt64 |
-| input_read_buffer_size         | 1048576    | 1048576    | SESSION | The size of buffer in bytes for input with format. By default, it is 1MB.                          | UInt64 |
-| max_block_size                 | 10000      | 10000      | SESSION | Maximum block size for reading                                                                     | UInt64 |
-| max_execute_time               | 0          | 0          | SESSION | The maximum query execution time. it means no limit if the value is zero. default value: 0         | UInt64 |
-| max_threads                    | 2          | 16         | SESSION | The maximum number of threads to execute the request. By default, it is determined automatically.  | UInt64 |
-| quoted_ident_case_sensitive    | 1          | 1          | SESSION | Case sensitivity of quoted identifiers, default value: 1 (aka case-sensitive)                      | UInt64 |
-| sql_dialect                    | PostgreSQL | PostgreSQL | SESSION | SQL dialect, support "PostgreSQL" and "MySQL", default value: "PostgreSQL"                         | String |
-| storage_read_buffer_size       | 1048576    | 1048576    | SESSION | The size of buffer in bytes for buffered reader of dal. By default, it is 1MB.                     | UInt64 |
-| timezone                       | UTC        | UTC        | SESSION | Timezone, default value: UTC,                                                                      | String |
-| unquoted_ident_case_sensitive  | 0          | 0          | SESSION | Case sensitivity of unquoted identifiers, default value: 0 (aka case-insensitive)                  | UInt64 |
-| wait_for_async_insert          | 1          | 1          | SESSION | Whether the client wait for the reply of async insert, default value: 1                            | UInt64 |
-| wait_for_async_insert_timeout  | 100        | 100        | SESSION | The timeout in seconds for waiting for processing of async insert, default value: 100              | UInt64 |
-+--------------------------------+------------+------------+---------+----------------------------------------------------------------------------------------------------+--------+
-
