--- src/query/service/tests/it/storages/system/users_table.rs
@@ -1,86 +1,0 @@
-// Copyright 2021 Datafuse Labs.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-use common_base::base::tokio;
-use common_exception::Result;
-use common_meta_types::AuthInfo;
-use common_meta_types::AuthType;
-use common_meta_types::UserGrantSet;
-use common_meta_types::UserInfo;
-use common_meta_types::UserOption;
-use common_meta_types::UserQuota;
-use common_users::UserApiProvider;
-use databend_query::sessions::TableContext;
-use databend_query::storages::system::UsersTable;
-use databend_query::storages::TableStreamReadWrap;
-use databend_query::storages::ToReadDataSourcePlan;
-use futures::TryStreamExt;
-use pretty_assertions::assert_eq;
-
-#[tokio::test(flavor = "multi_thread", worker_threads = 1)]
-async fn test_users_table() -> Result<()> {
-    let (_guard, ctx) = crate::tests::create_query_context().await?;
-    let tenant = ctx.get_tenant();
-    ctx.get_settings().set_max_threads(2)?;
-    let auth_data = AuthInfo::None;
-    UserApiProvider::instance()
-        .add_user(
-            &tenant,
-            UserInfo {
-                auth_info: auth_data,
-                name: "test".to_string(),
-                hostname: "localhost".to_string(),
-                grants: UserGrantSet::empty(),
-                quota: UserQuota::no_limit(),
-                option: UserOption::default(),
-            },
-            false,
-        )
-        .await?;
-    let auth_data = AuthInfo::new(AuthType::Sha256Password, &Some("123456789".to_string()));
-    assert!(auth_data.is_ok());
-    UserApiProvider::instance()
-        .add_user(
-            &tenant,
-            UserInfo {
-                auth_info: auth_data.unwrap(),
-                name: "test1".to_string(),
-                hostname: "%".to_string(),
-                grants: UserGrantSet::empty(),
-                quota: UserQuota::no_limit(),
-                option: UserOption::default().with_default_role(Some("role1".to_string())),
-            },
-            false,
-        )
-        .await?;
-
-    let table = UsersTable::create(1);
-    let source_plan = table.read_plan(ctx.clone(), None).await?;
-
-    let stream = table.read(ctx, &source_plan).await?;
-    let result = stream.try_collect::<Vec<_>>().await?;
-    let block = &result[0];
-    assert_eq!(block.num_columns(), 5);
-
-    let expected = vec![
-        "+-------+-----------+-----------------+------------------------------------------------------------------+--------------+",
-        "| name  | hostname  | auth_type       | auth_string                                                      | default_role |",
-        "+-------+-----------+-----------------+------------------------------------------------------------------+--------------+",
-        "| test  | localhost | no_password     |                                                                  |              |",
-        "| test1 | %         | sha256_password | 15e2b0d3c33891ebb0f1ef609ec419420c20e320ce94c65fbc8c3312448eb225 | role1        |",
-        "+-------+-----------+-----------------+------------------------------------------------------------------+--------------+",
-    ];
-    common_datablocks::assert_blocks_sorted_eq(expected, result.as_slice());
-    Ok(())
-}
